Should fix registered user count
This graph should be renamed to be "Total Registered Users" as "User::count()" is the total number of entries in the "user" table. Unless there is a way you can work out if users are actually logged on. 
ie. if there is a session timeout, then a date based token can be used, then a call to "User::active()" could return only those with token dates less then session timeout.
Just an idea.